+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [Add action to bulk categorize logs #590](https://github.com/farmOS/farmOS/pull/590)
+- [Add action for bulk assigning asset parent #588](https://github.com/farmOS/farmOS/pull/588)
 
 ### Changed